backend/src/main/java/kr/co/yigil/member/util/MemberUtils.java
package kr.co.yigil.member.util;

import kr.co.yigil.global.exception.BadRequestException;
import kr.co.yigil.global.exception.ExceptionCode;
import kr.co.yigil.member.domain.Member;
import kr.co.yigil.member.domain.repository.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

@Component
@RequiredArgsConstructor
public class MemberUtils {
    private final MemberRepository memberRepository;

    public Member findMemberById(Long memberId) {
        return memberRepository.findById(memberId).orElseThrow(
            () -> new BadRequestException(ExceptionCode.NOT_FOUND_MEMBER_ID)
        );
    }
}
